The "powerpc" port which has supported the newer Apple Macintosh powerpc based is being renamed to macppc. This is to allow sharing of common code between different powerpc base platforms.
Most of the work involved in the renaming process was performed by miod@

Files moved from powerpc/pci to macppc/pci

The file pci_machdep.h was not moved in this conversion.
It needs to be check if it is correct that should be the only shared
powerpc/pci file. Or if that file too should be MD, or more files MI.