Replace unreadable snprintf() and pointer arithmetic with simple strlcat().
If group file line would grow to be too long, leave it unmolested (previously
it would get removed).

Use fclose(fp) not close(fd) for a stream that was fdopen()ed.  Otherwise
we could leak memory.

When calling fwrite() pass the buffer as a single element instead of
using n one-byte elements.