Fix bsort() so that trailing spaces are not emitted.
When a system is rebooted rather than power cycled it is possible
for duplicate entries for a disk or cd to appear in the dmesg
output. One from the previous boot and one from the current
reboot.

bsort() correctly filtered out the duplicates, but could emit a
trailing space.

This trailing space caused problems when, for instance, the root
disk was being determined. With the trailing space present, the
install or upgrade script would not fill in the default root disk,
even if the system only had one disk.