pullup to 3.0:
Don't call adjust_linklocal() when creating INET6 sockaddrs, and
recognise IPv6 host routes as not requiring a netmask when
writing to the routing table socket.
This gets rid of the warning messages about failing to delete
routes at ppp startup and exit.

Ignore sockaddrs with sa_len of 0 returned by the PF_ROUTE/
NET_RT_DUMP sysctl.  The default route's netmask sockaddr always
has a sa_len of zero.

Ensure that when ppp stores a range with an IP address of 0.0.0.0,
it always has a netmask of 0.0.0.0 too.
This prevents ppp from mis-reading the default route as 0.0.0.0/32.

Pad sockaddrs written to the routing socket to sizeof(long).
This is a no-op on all archs for sockaddr_in and sockaddr_in6.