New feature request by (most recently) drahn@:
Do not allow the same mountpoint to be specified for >1 disk
partition.

If a mountpoint that has already been specified is entered again an
error message of the form:

Invalid response: sd0a is already being mounted at /var

is emitted. Partitions on both the disk being initialized and
previously initialized disks are checked.

Rework a bit of the logic to simplify and shorten in compensation. e.g.
eliminate references to /tmp/fstab.<disk> as it is never created.