Fix two problems:
1) When asked to choose a device, a user could enter two identical
words separated by a blank (e.g. 'a a') and thereby confuse the check
for selection validity. Quote user response before using it.

2) When asked to choose a sub-timezone, the user could escape to
a shell and when returning only see half the sub-timezone prompt
redisplayed. Put entire prompt into ask() parameter rather than
displaying first half in a separate echo -n.

Also change a "(or done)" to "(or 'done')" to be consistant with
other similar displays and move the initial comment for mount_a_disk()
to before the function to conform to normal usage.