from marc:
brad@ reported that the libpthread execve regression test failed on
alpha and macppc.   The test was somewhat bogus and depended upon the
compiler picking the global when a global and local were named the
same.   Anyway, this patch should fix it.  Tested on Henning's alpha
(bet he forgot he gave me an account :-), i386, and sparc64.