Introduce a new target switch, -mmemcpy (and -mno-memcpy).
This switch forces block copy operations to always fallback to memcpy()
when it is not done by inline load and stores, rather than using the
m88k movstr* functions found in libgcc.

The name of this option has been choosen after the mips back end, which
has a similar option.

Right now, this is a no-op since these functions are disabled due to (yet)
another gcc bug; but this will help building standalone code, such as the
kernel and boot blocks, without having to link against libgcc.