MFC: Fix by millert@
Traditionally, getopt(3) has treated "--foo" the same as "--".  However,
this can cause confusion when a user tries to use a long option with
a program that only supports short options.  Furthermore, it appears
to be in violation of POSIX, which states that "--" shall indicate
the end of argument processing, not any string that begins with "--".
OK otto@ and closes PR 3666.

ok deraadt@ millert@