make package opening more efficient (and maybe more intuitive) by delaying reading the full info directory to a temporary directory.
When we open a package, we just scan for the CONTENTS. If it's the
first file, we stash it into memory and STOP there, delaying grabbing
the rest of the special files to the first call to info().

Make sure the next() proxy shows only real file, by grabbing info files
if necessary.

Introduce handle->plist() method that can retrieve the plist from either
the version stored in memory, or in a file in the temp directory.