MFC: Fix by aaron@
In hardclock(), do not setup process virtual and profile timers if the process
is in the midst of exiting.  This solves a race condition that causes freed
memory to be left referenced in the master kernel timeout worklist, leading to
a uvm_fault (observed on an i386 MP system).

ok deraadt@ aaron@