Deja-vu - again I make bge stop claiming all interrupts for it's own.
This appears to have been due to driver changes and the fact that
the status word seems to not be cleared. linux may have the same bug.
freebsd does clear the status word with and atomic read and clear
ops. we instead do best effort here to check for the interrupt being
ours by reading the status from the status block to a local copy,
then clearing the status word once we enter the interrupt handler
and using the local copy for checking link status

ok jason@