vmware emulates mpi, but it does a half arsed job of it. half the fields we read off the hardware and use to configure the driver with are set to zero, so things dont really work like we want them to.
one of these fields is the pci subsystem id which is something we can fetch
really early in the attach process. so if the subsys is 0 then we go on and
fix up some of the values we get off the "hardware". now we can attach
disks on vmware.

"sneaky" and ok marco@ tested by and ok brad@