Finally fix the EEPROM reading code so that we can access the EEPROMs on all devices. In order to access the EEPROM, we must select 'EEPROM programming' mode, and then set the EEPROM chip select bit. Previously, we were setting both bits simultaneously, which doesn't work: they must be set in the right sequence.
Always obtain the station address from the EEPROM, now that EEPROM
reading works correctly.

Have re_tick() check for link state and set rl_link appropriately, if
there is a link then also check if the send queue is empty and if
not then try to send any queued packets.

Have re_start() check for the lack of link state or if OACTIVE is set,
if so then return.

From FreeBSD