Change the way the nmea(4) line discipline takes the timestamp for the timedelta sensor when no PPS signal is available.
Previously, the timestamp was taken at the leading '$' character
of a GPRMC message, which was not always correct, as some GPS units
send other messages first; we do not know when the GPRMC message
is sent within a data block (we refer to a data block as the set
of NMEA messages that are sent by a GPS unit in one go, usually
once per second).

nmea(4) now takes the timestamp at the first '$' character received
after the start of a new seconds.  Since GPS units transmit a data
block every second, the first message can be found by measuring the
gap between consecutive messages:  after the longest gap, the first
message of the next second follows.  And it is at the leading '$'
character of this message that we take the timestamp.

$GPGGA,.....    <- take timestamp here
$GPGSA,.....
$GPRMC,.....    <- decode time here
...             <- possibly more messages
		<- longer gap till start of next second
$GPGGA,.....	<- it starts all over, take next timestamp

This code has been designed to work independent of the baudrate and
the rate at which the GPS sends out it's data blocks (usually 1 Hz,
but 5 Hz units are common as well):

With this change, precision is greatly improved in the absence of a
PPS signal and as a side effect, jitter is reduced.

Note that while this is much better than before, there is still a
slight offset to the real time, as calculating the fix in the GPS unit
and transmitting the '$' character takes a short, but unpredictable
amount of time.

When tty timestamping is enabled, but there is no PPS signal available,
the sensor status will be degraded to CRITICAL, which means "check your
hardware".

Thanks to Chris Kuethe for testing and feedback.  ok ckuethe.