Change API for struct ata_xfer allocation to move it into the device that atascsi will be driving, and add it to the AHCI CCB.
This effectively gives us all the resources we need for a transfer in one
hit, meaning that we don't need to worry about whether we will have a pool
shortage or not enough CCBs.  The SCSI mid layer should take care to never
exceed the number of CCBs we have available, based on our sc_link.openings.