Switch the vax pmap to __HAVE_VM_PAGE_MD.
pv_entry head of lists are no longer preallocated at boot, and will only be
allocated for managed pages (instead of all physical memory pages, including
those containing the kernel).

pmap and pv_entry will now be allocated from a pool, instead of malloc for the
former and a homegrown allocator which never relinquishes unused elements to
the VM system for the latter.

The net result is a slight decrease in memory usage, and better behaviour in
low-memory conditions.