Keep running the scancode translation state machine even when the keyboard is set to raw mode, but only feed decoded events to wscons if non-raw mode.
Then, as an exception to this, pass the audio control keys events to
wscons even if in raw mode, so that the in-kernel mixer control code can
perform its work. The event is also seen by the raw event consumer (i.e.
the X server).

Requested and tested by many...