fix serial console handling on amd64. currently only glass console ever works, even if the boot loader is configured to use serial console.
after the changes jsing and i made, the initial probe of the serial port
was failing since the default base addres for the port is invalid. cos of
that the kernel would only ever consider the glass console for use. this
fixes it by reprobing the serial port if we get any config for it from the
boot loader.

advice on aesthetics from miod@
advice and ok from kettenis@