If you have #if FOO > 0 parts of your softc be sure to #include "foo.h" in all code that uses the softc, otherwise some object files will have a different idea of how big the softc struct is.
This will result in things like part of the softc, say the interrupt
handler being overridden by other data.  Putting printfs in the
main object file will look fine as it has the right idea as to
what the size is.  It is only after several hours debugging
that you start to figure out what the hell is going on.

jsing started trying to figure out the problem and got hints
from kettenis that it likely wasn't an interrupt problem,
after quite a bit of head scratching with joel I eventually spotted
the problem.

ok reyk@ jsing@