Don't be so pessimistic about the prospects of getting a valid MAC address.
Pyun YongHyeon seems to have taken code for generating an
address when something looks invalid from other drivers
in FreeBSD, there are no known problems with getting
valid addresses for jme, so remove this.

Additionally it isn't appropriate to call arc4random() at
that point in the boot process so the code needs to go.