now that some devices will get output from/to two different sources/ outputs by default it's nice to have a way to control multiple sources/ outputs with a single control setting.
so here's an implementation of "master" controls that uses selectable
control lists, because the biggest problem is figuring out what should
be "master".

outputs.master.slaves is the current set of output mixer controls
that settings for outputs.master and outputs.master will be applied
to.  record.volume.slaves works similarly for recording inputs.