Make vertex_free() correctly free LSAs in per-interface LSA trees.
Save a pointer to the right LSA tree in new member v->lsa_tree.
This saves us the hassle of finding the right tree in vertex_free(),
we already know it at creation time so why not tuck it in there?

This might cause problems if the ospfe side ever decides to call
area_del() with a non-empty area->lsa_tree, without correctly
initialising v->lsa_tree. But grep shows that the area's lsa_tree
is currently just initialized on the ospfe side, but never modified.

ospf6d daemons which crashed after about 1 min with malloc options
FGJ are now happy.

ok claudio@