In ospfe, originate new Router and Network LSAs before informing RDE about nbr state change.
Intra-Area-Prefix LSAs refer to these LSAs. While not strictly necessary,
it makes sense for these LSAs to be up-to-date in the LSDB before new
Intra-Area-Prefix LSAs are originated.

This change has no effect yet because Intra-Area-Prefix LSAs are
currently originated during updates of LSA types it depends on,
but origination will soon be synced to state changes in the iface
and nbr state machines instead.

ok claudio@