rework the polling code to use the semantic krw@ proposed.
intercept the ccb_done handling so polled commands set a flag that mpi_poll
tests on. when ccb_done sets the variable, the poll loop breaks and
mpi_poll runs the original ccb_done handler for the ccb completion.

this is a lot simpler than the previous implementation and removes a
mutex.

ok beck@