Work around yet another MC88110 errata, causing cache writebacks upon snooping to sometimes fail, by only using cache operation sequences known not to trigger the errata, at the expense of a small (barely noticeable) performance hit.
This workaround is only enabled when compiling a MULTIPROCESSOR kernel, since
the problem will not arise when using only one processor.

This change, refined over a few months, finally makes GENERIC.MP run stably
on 197DP systems. Of course, just writing this is enough to invoke Murphy's
law and make it fail again in an even more obscure and rare way...