Bad tedu, no cookie.
Don't set SDEAD on the process in exit1 untile we have grabbed the
allproclk.  allproclk is a rwlock and thus we may sleep to grab hold of
it. This is a big of a bugger when we just set a flag that means we
panic if we sleep.

ok art@. turns Tom Murphy's fstat panic into a deadlock instead *sigh*,
this is being looked into.