MFC, fixing failures during savecore (in the boot process) on some systems. Spotted by Mitja Muzenic and deraadt@ on various Dells. ok deraadt@
---
revision 1.46
date: 2010/12/02 20:57:07;  author: deraadt;  state: Exp;  lines: +5 -5
After the most recent change, make it possible to make -j again.  The
early MD and late MI files must be split up so that vers.o can sneak
between.  Issue spotted by bluhm, repair discussed with miod
---
revision 1.45
date: 2010/12/02 04:35:01;  author: deraadt;  state: Exp;  lines: +4 -4
move vers.o to before the other objects, so that it is not linked last.
having it linked last is bad (on at least i386 and amd64) because the lapic
is mapped over the start of the data segment -- savecore(8) then reads the
version string for a fixed buffer space, and reads into the lapic area
causing unintended side-effects (at least on Intel X5570 and X5680)
found by pedro, discussed with kettenis and mpf and miod
---