Document what remains of the 2BSD heritage. All facts found on the CSRG CD 1 in the 2bsd directory. Feedback and ok jmc@, ok sobrado@.
Note the large number of 3BSD -> 2BSD changes.
Kirk McKusick asked Cynthia Livingston (cael@, of USENIX, who converted
most manuals from man(7) to mdoc(7) for the 4.3BSD-Net/2 release and
added lots of historical information around that time) what happened
in this respect, and she answered that she often wrote "appeared
in 3BSD" when she was unsure of when a utility was added.
Thanks to Kirk and Cynthia for that clarification.