- use Lk for links
because i already have a trivial addition in afterboot, to briefly document
rc.d(8), i have committed that bit also. that saves me making (even more)
trivial changes to this file, however the text was suggested by ingo and
i'm still waiting for his reply which, to be fair, has only been a wait of
some 30 mins so far. so, the rc.d blurb was requested by ross richardson,
suggested by ingo, and probably ok ingo. but if he objects, we can tweak
it further.