Implement a simple wait queue API.  The idea is to allow multiple "waiters" to wait on the same "tag" for a deferred result.
A waiter is a callback and a void *argument.  The first waiter (the one for
which waitq_wait() returns true) is supposed to run some code that leads to
waitq_run() being run, which will destroy that waitq and call all callbacks
in turn.

Not used at the moment, but will be soon.

ok gilles@ chl@