Rework of the suspend/resume logic.
Do not places the ports into suspend mode now that we detach/attach
all USB devicess during a suspend/resume cycle.

Emit a reset when reinitializing the controller after resume and
try to use the same logic as in ehci_init().

Correctly wait for the controller to set the HCHalted bit to one
before attempting a reset.

Tested by many and have been in snap for a week.