Do no try to clear the endpoint, as it if has stalled, if the USB transfer reported an I/O error.
This is more likely to have no effect apart generating more
errors because when this happens, the device is generally
already physically detached but the *detach() task has not
run yet.  So simply return and wait for the device to be
removed like ucom(4) does.

Partially fix an issue reported by RD Thrush rt at thrush.com.

ok miod@