Resolve the uncertainty in the REFID assignment.
Previously, when there is an even number of offsets, we did the average
of the two middle offets but would set the REFID from one of them.
Instead, we simply select the middle offset with the lowest delay.

diff from Mike Miller <mmiller mgm51 com> (many thanks!)

OK phessler@, henning@