Fix library search order in our libtool.
Before, libtool first searched all directories for .la files, and if
search failed, switched to actual libraries (.so/.a). But the correct way
is to check each directory first for .la, then for .so/.a.

The problem was reported by ajacoutot@ who verified that the patch fixes
his case and okay'ed the initial diff. And jasper@'s bulk uncovered a bug
which is fixed now (verified by another bulk).