Fix hang with virtio event_idx feature
When using the RING_EVENT_IDX feature, we must first call publish_avail_idx()
and then read VQ_AVAIL_EVENT(vq), or there is a race condition that may cause
us to miss that the host needs to be notified.  This resulted in an occasional
hang of network in vio(4).