i broke the userland shim used for the extent regress test when i made it so struct pool was only visible to _KERNEL. tedu broke it too when he added the size argument to the kernel free functions.
this fixes both issues. the main change is to provide a local version of
struct pool with just the bit (pr_size) needed for extent to run.

if extents take advantage of more malloc/pool features (eg, {M,PR}_ZERO
then this will need to be updated again.

found by and based on a diff from Theo Buehler
ok mpi@