When a local route entry is added for an ifa having a broadcast address, also adds a broadcast entry flagged with RTF_BROADCAST.
Prior to this change broadcast entries were simple clonned ARP entries,
that would be deleted once their timer expired since they would always
be incomplete.

With this change they are now persistant and identifiable with a new flag.

Committing early to be able to deal with any potential fallout before we
start relying on this.

ok florian@, mikeb@, henning@