Add iwm(4), a new driver for Intel 7260 wifi cards.
Based on iwn(4) and Linux iwlwifi (which is dual BSD/GPLv2 licenced).
Created by Fixup Software Ltd. for genua mbh, who then passed on the code
base to the OpenBSD project.

The genua version of this driver was written for OpenBSD 5.4; ported to
-current by myself, phessler@, and deraadt@. So far, we've done semantic
and stylistic cleanup without functional changes. The driver is functional
but has some known issues which will be worked on in-tree.

Requires iwm firmware which is available in ports thanks to sthen@.

ok deraadt@ phessler@