Shrink the I/O map from 64KB to 16KB. This map is only used for smaller-than-a-logical-page allocations, while logical page size and larger are passed to uvm.
So in the worst case, the kernel will end up needing about 20 vax pages out
of it: 1 for non-console serial chips, up to 2 per Ethernet controller,
1 per SCSI controller, up to 3 for the clock or SSC chip, 1 for the cpmbox,
up to 3 for model-specific mappings (VS_REGS, KA650 stuff), 1 for the leds,
and up to 3 for frame buffer registers.

Thus shrinking from 128 pages to 32 is still large enough.