reintroduce page item cache colouring.
if you're having trouble understanding what this helps, imagine
your cpus caches are a hash table. by moving the base address of
items around (colouring them), you give it more bits to hash with.
in turn that makes it less likely that you will overflow buckets
in your hash. i mean cache.

it was inadvertantly removed in my churn of this subsystem, but as
tedu has said on this issue:

> The history of pool is filled with features getting trimmed because they
> seemed unnecessary or in the way, only to later discover how important they
> were. Having slowly learned that lesson, I think our default should be "if
> bonwick says do it, we do it" until proven otherwise.

until proven otherwise we can keep the functionality, especially
as the code cost is minimal.

ok many including tedu@ guenther@ deraadt@ millert@