Fix re-associations with athn(4) USB devices in station mode.
Sending the firmware a NODE_ADD command for our BSS doesn't seem to have the
desired effect in station mode. It made a new association after ifconfig
down/up impossible until the device was re-plugged. Leave the code in place
but limit it to hostap mode, just in case it is needed there (hard to tell
since hostap mode seems to be broken -- no beacons are sent by the device).

Also, when leaving run state, reset the RX filer to allow all beacons again.

ok mpi@