Change the default behavior of the s_time command so that it will perform a proper shutdown by sending a "close notify" alert to the server.  This allows s_time to benchmark a full TLS connection more accurately.
Introduce a new flag called -no_shutdown to make s_time adopt the
previous behavior (i.e. shut down the connection without notifying the
server) so that comparisons can still be made with OpenSSL's version.

The idea of using a flag (which replaces a #define) was suggested by
bcook@.  Thanks to millert@ and miod@ as well for their feedback on an
earlier diff which resulted in this change.

ok bcook@ beck@