The file(1) magic-parsing process was using pledge "stdio getpw proc recvfd" early on, then a set of getpwnam/setresuid/... before quickly dropping to "stdio recvfd".  It receives fd's and runs the magic code on them in a chroot'd "stdio" jail.  We can do better than that.
Before the recent change, "proc" contained both the concepts of "forking"
and "setuid".  "id" is now split out as a seperate request, and it is
exactly what this process needs momentarily.  So this loses another window
of opportunity, in case we have a major bug in .... hmm, it'd have to be
in getpwnam....

ok tedu doug semarie gilles