Insert a hack to deal with interfaces removing the VLAN header before the packet has been feed to the pseudo-interfaces input handlers.
To fix that without introducing a layer violation we should be able to
disable HW-vlan on parent when in use with different pseudo-interfaces.

In the case of bridge(4) for example it makes no sense to let the interface
remove the VLAN header if the kernel has to add it back for every packet.

Fix issues reported by sebastia@ and markus@

From dlg@, ok claudio@