Do not shut down the session upon receiving unknown messages.
RFC 5036 says the following about the receipt of unknown messages:
  "Unknown message bit.  Upon receipt of an unknown message, if U is
  clear (=0), a notification is returned to the message originator;
  if U is set (=1), the unknown message is silently ignored".

We were correctly ignoring unknown messages when the U-bit was set. But
when this bit was not set, we were shutting down the session when the
correct thing to do is to just send a non-fatal notification message.

Fix IxANVL LDP test 22.13.