rework sparc64 splfoo functions to be more consistent with other archs
this also moves us toward having an MI splraise().

sparc64 (and sparc) are different to the other archs because they
have macros that build templates. each spl uses that macro to create
an instance of an inline function specific to that spl call.

this moves it to having a single splraise inline that the spl api is
defined with. eg, #define splfoo() _splraise(IPL_FOO).

ok kettenis@