Streamline and refactor the code a bit more:
There are two sentinels for infinite output: infinity == true and reps == 0.
Ensure that infinity is set to true whenever reps is set to zero, then we
can always use 'if (infinity)'.

This allows us to merge the loop for deterministic output into the first
if (!randomize) statement, which is a lot tidier.

Insert an occasional empty line.

"move fast and break stuff" tedu