Various improvements to make this work better.
Restrict the block size to 512 bytes for now.  While Linux and NetBSD seem
to allow larger block sizes (4096 and 8192 bytes) I'm getting errors wth an
older 2G Sandisk SD card that advertises support for 1024 byte blocks.

Implement switching between the 400 KHz "discovery" clock and the 25 MHz
"operational" clock.  Enable the right clock based on the SD/MMC "unit"
number.  Don't advertise highspeed capabilities for now as I'm not sure
how the adjust the clock timing in that case.

Fix the way we calculate the number of blocks and the size of the total data
transfer to matc what we do in sdhc(4).