Reevaluate received mappings upon receipt of address message.
If a neighbor was a nexthop for a given set of prefixes but it's not
anymore, then we should uninstall the associated label mappings (if any)
from the kernel.  The same applies for the other way round (neighbor
wasn't a nexthop for a given set of prefixes but now is).

This issue is only evident when we have multiple links between a pair
of LSRs. Generally, when a link is shut down, the whole LDP is torn down
and all label mappings uninstalled automatically.

ok claudio@