Remove _ds_boundary abuse (again)
The logic behind this change is this: a single mbuf may reference
only a contiguous chunk of memory.  When this chunk crosses a page
boundary only the first part of it has a non-zero offset while all
other chunks start at the beginning of the page.

We take advantage of this fact and calculate the offset of a first
chunk as a simple "mtod(m, vaddr_t) & PAGE_MASK".