Using '4' as the max # of ccb's in a work unit doesn't work so well when the number of chunks in your RAID5 is significantly more than 4.  Each work unit needs to use at least a ccb per chunk to do the i/o.
Set the max to the number of chunks, which all the other RAID types
do in one varient or other. Note that it's not really a max, just
the number used when allocating the entire collection of ccb's for the
volume.

Fixes doing largeish i/o's (e.g. dd bs=1m count=128) to RAID5 volumes
with many chunks.

Problem reported by Alex McWhirter.

ok jsing@