refactor m_pullup a bit.
the most important change is that if the requested data is already
in the first mbuf in the chain, return quickly.

if that isnt true, the code will try to use the first mbuf to fit
the requested data.

if that isnt true, it will prepend an mbuf, and maybe a cluster,
to fit the requested data.

m_pullup will now try to maintain the alignment of the original
payload, even when prepending a new mbuf for it.

ok mikeb@