add poisoning of items on the per cpu caches.
it copies the existing pool code, except it works on pool_list
structures instead of pool_item structures.

after this id like to poison the words used by the TAILQ_ENTRY in
the pool_list struct that arent used until a list of items is moved
into the global depot.