Add ability to change media type
Tested with a X540 interconnected with a X550 via a CAT6 twisted
pair cable, but is expected to work on multi-speed fiber modules
as well to select between 10GbaseLR and 1000baseLX or 10GbaseSR
and 1000baseSX, etc.

This is largely required because X550 doesn't provide support for
auto-negotiation and requires manual configuration.

Obtained from FreeBSD.