Add Copyright and license. Pull in a small number of typo fixes from OpenSSL. Drop speculations about future developments that OpenSSL deleted as well.
This page is a mess for at least three reasons: (1) it doesn't
actually document all functions listed in the SYNOPSIS, (2) it ought
to be split into functional groups of reasonable size, and (3) the
user interface is ill-designed to the point of absurdity.  In view
of item (3), i'm loath to tackle items (1) and (2), so i'm leaving
this in its present sorry state for now, refraining even from
renaming it to ENGINE_add(3) and expanding the NAME section as
OpenSSL did - because that would merely sweep all the turds under
the rug.