Stop exporting the eproto_db array, export a pointer to it instead.
tcpdump directly uses eproto_db even though it is not part of the
libpcap API.  This means that we can't freely add members to this array,
else ld.so complains about size mismatches.  Keep the data in a static
array instead and make it usable by tcpdump through a pointer whose size
won't change in the future.  A minor bump is enough here for ld.so to
stop complaining.

While here, mark _eproto_db and llc_db as const, as they are meant to
be.

Suggested by and ok deraadt@