s/SSL_CTX_need_tmp_rsa/SSL_CTX_need_tmp_RSA/ because that's what <openssl/ssl.h> #defines.
That's likely a typo in the header file because all the other
functions are called *tmp_rsa*().  But it would be a bad idea to
fix such a bug in interfaces that are only provided for backward
compatibility in the first place, so i'm adjusting the manual to
be bug-compatible with the code, for now.

But, pretty please, for the next major bump, somebody go get Bob's
flensing knife and excise this part of the interface.  Like, export
ciphers?  Really?