Switch the connection to ftp.openbsd.org from using the ip address to hostname as preparation for an upcoming change.
But to ensure proper name resolution in case dns is not available
yet, add an entry to the hosts file. After an installation, remove
it so it does not end up in the final hosts file. Not needed for
upgrades because the file is not copied to the upgraded system.

OK deraadt@