Complete rewrite: Better one-line description. Specify the correct header file. Same parameter names as in ASN1_item_d2i(3). Lots of new information.
The ASN1_OBJECT interfaces appear specifically designed to maximize
the number and subtlety of traps, maybe in order to trap the wary
along with the unwary.  All the quirks, caveats, and bugs of
ASN1_item_d2i(3) apply, and there are three additional ones on top
in this page.

It looks like that design approach was so successful that the designers
managed to trap even themselves: see the new BUGS section.