rework how we address physical disks on the second scsibus.
to talk to a physical disk on mfii you need to know its physical
disk id and its current dev handle.

you can fetch a list of physical disks with a MR_DCMD_PD_GET_LIST
command, and you get the dev handle by fetching the associated info
from a MR_DCMD_LD_MAP_GET_INFO command.

i made a mistake and thought that the index into the PD_GET_LIST
info was the target id of a disk, and that should also be used as
the index into the LD_MAP_GET_INFO structures. it turns out that
the PD_GET_LIST is a dense array of info that points at the target
id, and the target id is used as the index into the LD_MAP_GET_INFO.

it also turns out that the dev handles from LD_MAP_GET_INFO can
change at runtime, so we should update them all when the bus topology
changes.

this change fetches a list of dev handles via LD_MAP_GET_INFO up
front, and makes them availble to the passthru bus scsi_cmd handler
via SRP. in the future hotplug events will update the map concurrently
with disk accesses.

also get rid of the probe for pd disks since the id from PD_GET_LIST
and what the scsibus automatically probes are the same. we just let
the midlayre blindly probe all possible targets on the passthru
bus. the pd scsibus probe handler still fetches specific disk info
to ensure we only provide access to disks marked as PASSTHRU in the
firmware.