handle physical disk state changes.
more specificially we probe the disk if it goes from UNCONFIGURED_GOOD
to a SYSTEM disk, and detach it if goes from being a SYSTEM disk
to anything else.

this semantic comes from the lsi^Wavago code in the illumos mr_sas
driver. seems to work fine.

i think this covers all the ways a passthru disk can transition on
these boards.