The videocore portion of the raspberry pi which boots the arm cores and runs the mailbox interface knows about a MAC address that appears to be derived from a unique serial number along with the raspberry pi foundation oui.
It modifies the device tree when booting to store the MAC address
in /axi/usb/hub/ethernet/mac-address so fetch and use this value
for the integrated smsc(4) Ethernet.

A different smsc adapter plugged into one of the USB ports
probes later with a different unit number and skips this path.