Support for displaying flow queues alongside H-FSC
This (ab)uses the fact that node->qstats.data.period field in
hfsc_class_stats structure is at the same offset as the 'flows'
field in fqcodel_stats.

While here make use of a presently empty field "SCH" to display
the queue management policy (flow or fifo) which is not strictly
a scheduler, but it will hopefully become descriptive and useful
later. This distinguishes flow queues from the regular HFSC ones.

OK sthen, visa