regress/sbin/pfctl - use -n for pfcmd group
Passing in the -n which just results in the rules being parsed.

The definition for the pfctlcmd testing group in the Makefile is:
# pfcmd: test pfctl command line parsing

This test passeed only because pfctl didn't exit with an error
when it was asked to clear tables from a non existing anchor.

with the -n flag we still test if the anchor command can be combined
with -Fa but don't actually attempt to run the clearing code.

OK mikeb@, sashan@