In iwm(4), instead of scheduling a task which calls ieee80211_end_scan(), call ieee80211_end_scan() directly from interrupt context.
This extra task was already part of the original driver code from 2015 (but
with a workq instead of a task). Back then, the driver had to run two separate
scan commands in succession, for 2 GHz and then 5 GHz. Which is why a task
was used, since sending another command requires a sleepable context.

Nowadays, with our current firmware, a single scan command is sufficient
so there is no code path which needs to sleep when the scan ends.

ok mpi@