Restore "print" in ddb; add "pp[rint]" for pretty-printing
Changing the "print" command to use db_ctf_pprint_cmd() broke all documented
uses and "examine" does not allow printing variables, or individual registers.

For now it is better to leave the "print" command intact, and in sync with the
ddb(4) man page, but we still want "pp[rint]" to replace the "print" command
when it works better.

ok mpi@