Make sure iwn(4) firmware gets to see a beacon before we send an auth request. Otherwise, the Tx attempt can fail due to the firmware's built-in regulatory domain enforcement.
Because this code runs in interrrupt context we use a DELAY(). It's currently
set to 3 times the beacon interval. This is a bit long as far as DELAYs go,
but since this driver runs state transitions in interrupt context there is
no better way for now. Lower delays weren't reliable enough in my testing.

Makes association work reliably for me on 5GHz. My original plan was to only
add this hack for 5GHz but Holger Mikolon reported that it also helps on 2GHz
channels with regulatory restrictions (e.g. 13), so do it always.

ok mpi@