clean-up option handling slightly, some more to be done.
Add a -S option that computes signatures, so that we may remove
code from bsd.port.mk that does the exact same thing, and no longer
work since the addition of @version.

For now, -S is exclusive of other package creation options, as it exits
as soon as it computes what it needs.