Allow to boot into the existing bsd.rd and to download, verify and upgrade to the bsd.rd of the next release.
The installer detects this bsd.rd only upgrade if the answer to
the 'Server directory?' question points to the sets of the next
release.

It then limits the set selection to the bsd.rd file, verifies it
using the on-disk signify key of the next release and makes a
backup of the existing bsd.rd before installing it.

Idea from and OK deraadt@
OK benno@