Split sys_ptrace() by request type:  - control operations: trace_me, attach, detach, step, kill, continue.    Manipulate process relation/state or send a signal  - kernel-state get/set: thread list, event mask, trace state.    About the process and don't require target to be stopped, need copyin/out  - user-state get/set: memory, register, window cookie.    Often thread-specific, require target to be stopped, need copyin/out
sys_ptrace() changes to handle request checking, copyin/out to
kernel buffers with size check and zeroing, and dispatching to the
routines above for the real work.  This simplfies the permission checks
and copyin/out handling and will simplify lock handling in the future.

Inspired in part by FreeBSD.
ok mpi@ visa@