Detect touchpad reset announcements. On rare occasions, the touchpad is reset due to power failure and an announcement is transmitted as input by the device making it detectable. At this point, the device must be re-enabled to continue working since any configuration done prior to the reset is lost.
Detection is implemented using a watchdog and enabled for all protocols. So far,
the announcement has been observed on Synaptics by me and on Elantech v4 by
bru@. Extra care has been taken to handle potential false positives: when the
announcement byte sequence is part of a valid input packet.

with help and ok bru@, mpi@