The payloads are layered like onions, so you can validate one layer and then call the next one, which can then validate itself.  Thing is, most layers try to run validations on the upper layer, which is not useful and rather confusing.  This cleans it up.
First change is that the generic payload parser does not anymore pass
the length of the whole datagram, including all remaining payloads, but
passes only the length of the specific payload to the specific payload
parser.  Second change is that the payload validators don't check the
length of the upper layer, but only verify their own lengths.

Diff discussed with hshoexer@ and sthen@
Tested by sthen@