add support for point to multipoint GRE via mgre(4).
mgre is different to gre in that the tunnel only needs configuration
for a local address, and the address inside the tunnel is configured
on it is a subnet. other addresses on that subnet get mapped to a
tunnel endpoint and send there.

at the moment that mechanism is the routing table, using non-gateway
host routes with the tunnel endpoint as the gateway address.

the tunnel address configuration is ugly and confusing atm, but
should improve soon.

ok mpi@ who has some suggestions i can do in the tree.