Disallow "++minutes".
strtonum(3) accepts a single leading '+', so if we increment timearg we
allow input with two leading pluses.  If we don't increment, we still
have a valid input for strtonum(3).

While here, use errstr to say what was wrong with timearg.

Don't increase the range for offsets yet: it exposes segfaults elsewhere
in the program that need to be addressed.

ok millert@ tb@