Asseert that a pfkey or routing socket is referenced by a `fp' instead of calling sofree(), when its PCB is detached.
This is different from TCP which does not always detach `inpcb's from
sockets.  In the pfkey & routing case caling sofree() there is a noop
whereas for TCP it's needed to free closed connections.

Having fewer sofree() makes it easier to understand the code and move
the locks down.

ok visa@