Better handle automatic column width assignments in the presence of horizontal spans, by implementing a moderately difficult iterative algoritm.  The benefit is that spans containing long text no longer cause an excessive width of their starting column.
The result is likely not optimal, in particular in the presence
of many spans overlapping in complicated ways nor when spans
interact with equalizing or maximizing colums.  But i doubt the
practical usefulness of making this more complicated.

Issue originally reported in synaptics(4), which now looks better,
by tedu@ three years ago, and reminded by Pali Rohar this summer.