split off "local" input handling
this means the current mpls header will be passed along with the
mbuf for mpw to look at. right now this doesn't do anything, but
it will allow for implementation of RFC 6391 (flow aware transport)
and using the exp header for cos.

when mpe gets moved to adding an RTF_LOCAL route, this will be used
for cos and ttl handling.