rework mpe_input so it patches ipv4 cksum, and handles ipv6 too.
rather than check and recalculate the ipv4 checksum, we can update
the cksum incrementally. this is a bit faster, and means we'll see
more things on bpf.

also, peek at the first nibble so we can tell ipv6 from ipv4. i
consider them the same FEC, so either can be received now.