Clarify preposition
Any messages sent to the kernel are returned, and copies are sent to all
interested listeners.  The kernel will provide the process ID for the
sender [...]

Change the last "for" to "of", making it even clearer that the kernel tells
listeners about the sender's PID;  it does not provide anything back to the
sender in particular.

"Seems right" deraadt