basic support for dumping xfp info
xfp uses a different page layout compared to sfp. xfp uses info in
INF-8077, not SFF-8472, and by default provides the info at page
1, not 0.

myx(4) supports xfp modules, but doesn't support i2c writes so it
cannot change the page the chip is looking at. it appears that myx
defaults to page 1 for xfp, so if the first read of page 0 fails
with ENXIO, try to read page 1 instead.

a 10G-PCIE-8A-R myx with the only xfp module i could find lying
around looks like this:

	transceiver: XFP
	model: Intel Corp TXN181070850X1J rev B3
	serial: USC2048M95 date: 2005-04-13

im not sure if it's worth spending much more time on these though..