Switched min_heap to size_t to prevent integer overflows.
Also, as deraadt suggested, switched realloc to recallocarray to at least
prevent uninitialized memory to be used as pointers in case of other
programming errors. A proper solution (not using an array) needs more work.

This change occured in sync with upstream libevent 2.2.

with input by and ok bluhm, jca, tedu