Improve the interaction between efifb(4), inteldrm(4) and radeondrm(4) when we have a serial console by introducing the notion of a "primary" graphics device.  The primary graphics device is the one set up and used by firmware (BIOS, UEFI).
The goal is to make sure that wsdisplay0 and drm0 reliably attach to
the primary graphics device such that X works out of the box even
if you have multiple cards or if you are using a serial console.

This also fixes the situation where inteldrm(4) or radeondrm(4) would
take over the console on UEFI systems even if the kernel was booted
with a serial console.

ok jsg@