Make sure that the IMSG_CTL_RESET message is sent immediately.
This fixes an issue that might better be solved in imsg itself.

The problem is that IMSG_CTL_RESET does not include an fd while the
following messages (IMSG_CFG_RELAY and IMSG_CFG_RELAY_FD) do contain
fds.  If the receiver gets them in one buffer (via recvmsg), the first
fd might be wrongly associated to the IMSG_CTL_RESET message.  This is
theoretically taken care of by the imsg API, so it is either a bug in
relayd's API usage or in imsg itself.

"sure" claudio@ as a temporary fix.