Invalidate icache when changing mapping
On octeon, the instruction cache appears to fulfill instruction fetches
on icache hit even if the virtual address mapping no longer exists or
does not allow execution. Prevent running of lingering instructions
by invalidating the icache when changing a mapping that has previously
been executable.

This fix is somewhat rough and relies on the availability of the XI bit.