amd64: ensure %fs is loaded after final lgdt
Mark Patruck reported this issue on a new EPYC 7402P CPU and tracked down
a DragonflyBSD diff that fixed the issue. I solicited some feedback/advice
from Matt Dillon of Dragonfly and discussions between him, guenther@ and
myself led to this fix which appears to resolve the problem.

ok guenther@