avoid rxr races between rxfill from the interrupt and timeout paths
vmx "hardware" seems to be able to use rx descriptors as soon as
theyre filled in, which means filling the ring from a timeout can
run conccurently with an isr that's pulling stuff off the ring.
this is mostly a problem with the rxr accounting, so we serialise
updates to the alive counter by running rxfill in a mutex.

most of the investigation was done by claudio@ and mathieu@
an earlier version of this diff was tested hard by mathieu@ and was
ok@ claudio