Change Elf64 types and structs to match the System V ABI specification.
Change Elf64_Half from a uint32_t type to a uint16_t type.  The size of
structs do not change as the previous uses of Elf64_Half become
Elf64_Word (uint32_t).

Remove Elf64_Quarter as 16 bit values now use Elf64_Half.
Replace Elf_Byte use with unsigned char.
Change some uses of Elf64_Xword to Elf64_Addr (both uint64_t).

Corrects behaviour with libelf where the file size of a symtab section
came back as 26 not 24 as libelf determines this by the size of a the
involved types not the size of the struct in exec_elf.h.  As a result
libelf's elf_getdata() returned ELF_E_SECTION due to the size of the
section not being a multiple of the file size of the section type.
This occurred with the new runtime linker for radeon shaders in
Mesa 19.2.

Sync description of Elf32 and Elf64 types in elf(5) with FreeBSD and
adjust types in structs mentioned.

ok guenther@ deraadt@