Fix uvm_unmap_remove panic when tearing down VMs.
Type confusion when tearing down a vmm(4) VM's vm_map led UVM to interpret
the vm_map as a struct uvmspace, and occasionally fail. stefan@ pointed out
the root cause a while back but I never had time to look at this until now.

Syzkaller also found a reproducer for this, so the "Reported by" line
below is a clue to let it know to close that bug.

idea by, and ok stefan@
also ok anton@

Reported-by: syzbot+dbd6e2a7583e958e69f8@syzkaller.appspotmail.com