ciss(4): tsleep(9) -> tsleep_nsec(9)
"i" is a count of milliseconds.  We convert it to a starting interval
"ts" and determine our absolute timeout "end".  Then we loop through
until the poll is successful or "end" elapses.

This could be simplified with an absolute timeout interface for
tsleep(9) but alas, we don't have one yet.

ok krw@