unify the tun and tap output and read behaviour.
let tap use ether_output directly, and then cut back tun_output so
it does the same things that ether_output does. specifically, this
means tun_output now only prepends the packet with the "link" header,
and no longer runs BPF for outgoing packets. running BPF for tun
packets in output used to be needed because pipex used to get a
chance to steal the packet at this point, but you would still want
to see the packet in tcpdump output. now BPF is handled in tun_dev_read
for both tun and tap.