move to if_vinput() in tun_dev_write.
this means tun doesn't queue the packet on input for the network
stack to process, it's pushed through as part of the write into the
kernel.

discussed at length with claudio@ who agrees that avoiding a queue,
and charging the writing process with the work associated with the
packet, are both reasonable (good) things to do.