*sleep_nsec(9): log process name and pid when nsecs == 0
We included DIAGNOSTIC in *sleep_nsec(9) when they were first committed
to help us sniff out divison-to-zero bugs when converting *sleep(9)
callers to the new interfaces.

Recently we exposed the new interface to userland callers.  This has
yielded some warnings.

This diff adds a process name and pid to the warnings to help determine
the source of the zero-length sleeps.

ok mpi@