move the SMR_LIST_REMOVE and smr_barrier up in tun_clone_destroy.
without this the tun_softc is still available on the list for the
syscalls to get to, even though the device is dead and should no
longer be referenced. by leaving it in the list after the
refcnt_finalize, it was still be found and was used.

found by claudio@
jmatthew@ agrees with the change