actually set the link state down when the /dev entry is closed.
this means a route message is sent when the interface is closed and
goes down, but also causes another route message to be sent when
the interface comes up on the next open. this is important for
things like ospfd and the ospfd regress test because they want to
know when link comes up.

the regression was pointed out by bluhm, who also helped me isolate
the problem.