Replace atoi with strtonum for stricter and safer integer parsing.
-N ndots: 0, INT_MAX
The upper limit is a bit silly, everything bigger than a small number
will force domain names to be interpretet as relative.

-R retries INT_MIN, INT_MAX - 1
Specifically documented to accept negative numbers.

-W wait 0, INT_MAX
One could interpret the documentation as meaning that it accepts
negative numbers but that failes later on with an error message from
the timer code..

with deraadt, OK kn, input & OK tedu