pcppi(4), spkr(4): ticks -> milliseconds
In pcppi(4), convert from ticks to milliseconds.

While we're doing this, we can also convert spkr(4), too.

The spkr(4) conversion from ticks to milliseconds is trickier because
the driver is written with musical intervals (whole notes, quarter
notes, etc.), not the typical units (seconds, milliseconds, etc.) used
in most software.

I think the conversion is correct... but the code is a challenging
read, so it might be subtly incorrect.

ratchov@ intends to move spkr(4) into the attic sometime soon so I
doubt it matters much if I got it wrong.

Input from schwarze@, jsg@, and ratchov@.

Tested by schwarze@.

ok ratchov@