acpi(4): acpi_sleep(): tsleep(9) -> tsleep_nsec(9)
kettenis@ notes that ACPI integers are supposed to be unsigned, but
ours are signed.  This is a much deeper problem than what I'm looking
to change with this conversion.  To work around the issue for now we
can round the sleep interval up to 1 millisecond.

There is a very nasty-looking rwsleep(9) call elsewhere in
dev/acpi/dsdt.c that I'm going to leave as-is for now to avoid
breaking something.

With input from pirofti@, kettenis@, and mortimer@.

Tested by Lucas Raab, Moises Simon, mortimer@, and gkoehler@.

ok kettenis@