Work around FIFO_UNDERRUN (0x84) Tx errors being reported by iwn(4) firmware.
When this error occurs, it tends to occur multiple times in a row. Affected
frames are lost and never get transmitted, and traffic stalls for a while.

I don't understand what is causing this. I have found that it only occurs if
we ask the firmware to use its multi-rate retry table. If we send frames at
a fixed rate, it does not happen. So this can be used as a workaround until
the real problem is understood.

Since this error is particularly problematic with block ack, use a fixed Tx
rate for aggregation queues for now. All subframes contained in an aggregate
are always sent together at the same rate anyway.

ok mpi@