make taskq_barrier wait for pending tasks, not just the running tasks.
I wrote taskq_barrier with the behaviour described in the manpage:

 taskq_barrier() guarantees that any task that was running on the tq taskq
 when the barrier was called has finished by the time the barrier returns.

Note that it talks about the currently running task, not pending tasks.

It just so happens that the original implementation just used task_add
to put a condvar on the list and waited for it to run. Because task_add
uses TAILQ_INSERT_TAIL, you ended up waiting for all pending to work to
run too, not just the currently running task.

The new implementation took advantage of already holding the lock and
used TAILQ_INSERT_HEAD to put the barrier work at the front of the queue
so it would run next, which is closer to the stated behaviour.

Using the tail insert here restores the previous accidental behaviour.

jsg@ points out the following:

> The linux functions like flush_workqueue() we use this for in drm want
> to wait on all scheduled work not just currently running.
>
> ie a comment from one of the linux functions:
>
> /**
> * flush_workqueue - ensure that any scheduled work has run to completion.
> * @wq: workqueue to flush
> *
> * This function sleeps until all work items which were queued on entry
> * have finished execution, but it is not livelocked by new incoming ones.
> */
>
> our implementation of this in drm is
>
> void
> flush_workqueue(struct workqueue_struct *wq)
> {
>	if (cold)
>		return;
>
>	taskq_barrier((struct taskq *)wq);
> }

I don't think it's worth complicating the taskq API, so I'm just
going to make taskq_barrier wait for pending work too.

tested by tb@
ok jsg@