Load iwx(4) firmware just once.
On iwx(4) devices there is no two-step firmware loading process with an
INIT and RT image, as was the case with iwn(4) and iwm(4).

Loading the image twice leaves firmware in a semi-operational state with
problems such as fatal firmware errors and missing regulatory domain
update notifications.

Tested by jcs@, sven falempin, and myself.

ok patrick@