Make iwx(4) reset Tx byte count table slots for frames that are done.
This fixes a "fatal firmware error" at run-time.

Thanks again to Sara Sharon who provided the hint that the error trace
I was looking at indicated that the firmware was running into a problem
while trying to flush its Tx queues, and that this could be related to
the Tx byte count table not being maintained properly.

Tested by sven falempin and myself.