Prepare for newer iwx(4) firmware scan command versions.
Add declarations of "version 2" scan flags from iwlwifi and prepare the
iwx_scan_channel_cfg_umac struct for life beyond version 1.

None of this is needed yet. But we will need this at some point and I've
already written the diff, hoping it would prevent firmware errors (which
of course it didn't).