Remove some customizations from the perl build
This was originally done so that other things in-tree such as vi could link
to libperl and to support non-shared vax.

This brings us more in-line with upstream and makes things more understandable.

It now links libperl.so to libm, which some software needs, and stops building
and installing libperl.a.

OK bluhm@