add kstats for rx queues (ifiqs) and transmit queues (ifqs).
this means you can observe what the network stack is trying to do
when it's working with a nic driver that supports multiple rings.
a nic with only one set of rings still gets queues though, and this
still exports their stats.

here is a small example of what kstat(8) currently outputs for these
stats:

em0:0:rxq:0
         packets: 2292 packets
           bytes: 229846 bytes
          qdrops: 0 packets
          errors: 0 packets
            qlen: 0 packets
em0:0:txq:0
         packets: 1297 packets
           bytes: 193413 bytes
          qdrops: 0 packets
          errors: 0 packets
            qlen: 0 packets
         maxqlen: 511 packets
         oactive: false