add kstat support for reading hardware counters.
this replaces the existing counters implementation, which just
collected the stats in the softc, but didn't really provide a way
for a person to read them.

em counters get cleared on read. a lot of them are 32bit, so to
avoid overflow the counters are polled and the newly accumulated
values are added to some 64 bit counters in software.

tested by hrvoje popovski and SAITOH Masanobu
ok mpi@