Switch the default pager from "more -s" to "less".
POSIX explicitly allows using a different default pager if that
is documented.  The pager provided in the OpenBSD base system is
less(1).  It can merely be called as more(1) for compatibility.
Our man(1) implementation uses less(1) features that traditional
more(1) did not provide, in particular tagging.  Besides, as noted
by deraadt@, the user interface of less(1) is slightly more refined
and preferable over the user inferface of more(1).
This switch was originally suggested by Ian Ropers.

As explained by jmc@ and deraadt@, the -s flag was added a very
long time ago when an antique version of groff(1) had an annoying
bug in terminal output that would randomly display blank lines in
the middle of pages.  Clearly, -s has no longer been needed for
many years, so drop it from the default pager invocation.

OK deraadt@ jmc@ martijn@ job@