Add support for UTF-8 DISPLAY-HINTs with octet length. For now only sysContact is supported with SnmpAdminString, but more shall follow soon(tm).
Note that this will change output even in the C locale: Previously every
unprintable byte would be tranformed to a '.', with this diff a valid UTF-8
multibyte or unprintable character will be squished into a single dot to
give a better understanding of the intention of the original message.
Invalid bytes will now be printed as question marks.

Lot's of help and back and forth with schwarze@ who also kindly offered to
walk away when I mentioned looking through MIB files for more objects to
convert to this new code, which is understandable.

OK schwarze@