Get the default values for font height and width in wsfontload(8) using the WSDISPLAYIO_GETSCREENTYPE ioctl. This ensures that they always match the currently loaded font metrics.
Previously, wsfontload(8) hardcoded the default height and width values
for the font to be loaded as 12x22 when using framebuffer consoles, and
as 8x16 when in text mode. The 12x22 value wasn't correct in case we
felt back to the smaller 8x16 font for screen widths smaller than 960px,
and wasn't valid anymore since we replaced Gallant 12x22 by Spleen 12x24
on all architectures but sparc64.

OK jcs@, mpi@