-F requires at least 5 columns for folding, while a tab at start of line is 8 columns. This causes an infinite loop in the folding logic.
Fix this by implementing an escape hatch so that only a single fold can
occur. This does mean that a tab with a -F smaller or equal to 8 will
always print the newline escape backslash in the 8th column.

Bug reported by phillbush <at> cock <dot> li

Minor tweak and OK deraadt@