Switch use of '\0' to NULL
Fix a clang 10 warning about comparing a pointer to a null character. The
condition "if ((s = symnam[i]) == '\0')" used to be "if (s = symnam[i])"
and the incorrect spelling of NULL was chosen in a -Wall cleanup 19 years
ago (-r 1.6). Upstream uses a naked 0 instead of NULL, so does NetBSD.

ok martijn millert