amd64: add tsc_delay(), a delay(9) implementation based on the TSC
In preparation for running the lapic timer in oneshot mode on amd64 we
need a replacement for lapic_delay().

Using the lapic timer itself to implement delay(9) when the timer is
not running in periodic mode is complicated if not outright impossible.

Meanwhile, the i8254 provides our only other amd64 delay(9) implementation
and it is an extremely slow clock.  On my 2GHz machine, gettick() takes
~20 microseconds to complete *without* mutex contention.  On a VM it is
even slower, as you must exit the VM for each inb() and outb().

So, add tsc_delay() and use it when we have a constant/invariant TSC.
The TSC is a 64-bit "up-counter" so the implementation is simple.

Given how slow the i8254 is on modern machines, we may want to add an
HPET delay(9) implementation as a fallback for machines where the TSC
drifts.  The HPET itself is pretty slow, but not as slow as the i8254.

Discussed with kettenis@, Mike Larkin, and naddy@.

Tweaked by kettenis@.

ok kettenis@