miod notes T_USER needs to be a bit, it was not.
The "snowflake" uniqueness of every MD trap impl often gets in the way
of precisely & correctly interfacing to MI layers. The differences
also complicates review, and causes new MI requirements to be
incorrectly written. Thus an architecture will fall behind, not just
because they are slow or rare, but because the code behaviour becomes
increasingly incorrect.  It is sad.