Test DTLS timeouts and retransmissions by dropping specific messages.
Provide a BIO that can drop specific messages in order to trigger and test
DTLS timeouts and retransmissions. Note that the SSL buffering BIO (bbio)
has to be removed to ensure that handshake messages are sent individually.

This would have detected the recent DTLS breakage with retransmissions for
a flight that includes a CCS.