mbg(4): more tsleep(9) -> tsleep_nsec(9)
The mbg(4) driver sometimes needs to spin until the MBG_BUSY flag goes
away.  If the kernel is cold it calls delay(9) for 20us an iteration
for up to 50 iterations.  If the kernel is not cold it sleeps for up 1
tick per iteration for up to hz/10 iterations.

To switch from tsleep(9) to tsleep_nsec(9) claudio@ suggests blocking
for at least 1ms per iteration for up to 10 iterations.  He reasons
that the flag change is expected to take ~70us, so 10 x 1ms (or more)
is more than enough.

Discussed with mpi@.

ok claudio@