when setting a flowid, set the M_FLOWID csum_flags bit too.
this "fixes" TCP going over an interface with fq codel enabled. the
way the codel code classifies a packet without a flowid set is to
randomly assign it to a bucket. this in turn means that packets
will get reordered, and tcp hates that.

sthen was able to find a test case and narrow down at which time
the problem appeared, helped greatly.

tested by sthen@ and millert@
ok sashan@ jmatthew@