Fix SSL_get{,_peer}_finished() with TLSv1.3
As reported by Steffen Ullrich and bluhm, the Finished tests in
p5-Net-SSLeay's t/local/43_misc_functions.t broke with with TLSv1.3.
The reason for this is that we don't copy the MDs over to the SSL, so
the API functions can't retrieve them.  This commit fixes this part of
the test (one unrelated test still fails).

ok inoguchi jsing