turn hardware rx mbuf timestamping off by default.
we see too many ntp replies (appear to) arrive before the request
for them was sent, which ntpd handles by disabling the peer for an
hour.

this was a lot easier to narrow down after fixing up bpf and
timestamps, cos it let me see this:

16:50:36.051696 802.1Q vid 871 pri 3 192.0.2.55.47079 > 162.159.200.123.123: v4 client strat 0 poll 0 prec 0 [tos 0x10]
16:50:36.047201 802.1Q vid 871 pri 3 162.159.200.123.123 > 192.0.2.55.47079: v4 server strat 3 poll 0 prec -25 (DF)

im going to borrow the link0 flag for a bit to allow turning
timestamping on.