rename filenames ending in a period
While windows may technically allow filenames to end in a period, the
recommendation is to avoid this as the shell does not support such files.
As of git 2.25, the behaviour was apparently changed to not allow files
with names ending in a period to be checked out on windows.

A second problem is upper/lower case conflicts for filenames.

To allow the source tree to be checked out on windows we follow FreeBSD
and rename a few files:

ss.. -> ss_
ssA  -> ss10
ssB  -> ss11

(FreeBSD svn revision 210167 from 2010)

as usual, this issue was extensively researched by jsg@