rename pkg_add's pod files to avoid filenames with colons
Filenames with colons are not allowed on windows filesystems. So we rename
the pod files to avoid this. Should get us closer to being able to check
out the openbsd source code on a windows system.

ok espie@ (who suggested the current approach); sthen@ was ok with a
previous version of the diff