expose the mcx timer as a timecounter.
this is mostly to help me better understand where i accumulate error
when trying to sync the chip to the kernel clocks. ie, if im using
mcx as the kernel clock source and my attempts to sync to it still
produce errors, then my code is very wrong instead of slightly
wrong.

it's also fun and a tiny amount of code.