- grab fullpkgpath early (to decide: hard error if not provided) ? - tweak the look-up mechanism for dependencies to NOT look at installed packages for inter-dependencies between subpackages (and debug packages) of the SAME pkgpath.
The rationale for looking at installed packages in the usual case
is that it's normally much faster rather than have the ports tree
spew out packing-lists.

In the case of multi-packages updates, the new version may have been built
successfully (and correctly) while the old version is still around.
However, if pkg_create looks at the installed version, it may fail to create
the package because the shared library versions will be different.