dhclient(8): fork_privchld, take_charge, propose_release: poll(2) -> ppoll(2)
Switch from poll(2) to ppoll(2) in a few more functions.

Because we're working with ppoll(2) and clock_gettime(2) it is easier
to encode the various timeouts as static const timespecs instead of
preprocessor macros.  This way we aren't packing timespecs in the
middle of the code, which distracts from the (more important) logic of
what the code is doing.

Part of a larger campaign improve "time stuff" in dhclient(8).

Prompted by and discussed with krw@.  Based on a diff by krw@.

ok krw@