Adjust unveil_find_cover() to return -1 if the root vnode is passed in. This helps unveil_add_vnode() to properly re-evaluate unveils when "/" is added to the list.
Because of this adjust unveil_covered() to check for the root as well
so that in that case the unveil uv is returned instead of NULL. Traversing
up from the root returns the root. This check is not really needed since
namei has its own root check and shortcuts for root vnodes.

OK semarie@