Run more device init code when resuming iwm(4) and iwx(4) devices.
deraadt@ pointed out that the resume code path only ran a small part
of the entire hardware init sequence which runs when the device is
first attached at boot time.

In particular, we didn't wait for device stabilization, and MSI-X
initialization was attempted too early, and the "persistence bit"
workaround was skipped (which, ironically, is supposed to address
a known hardware bug during resume).

Tested by jcs on AX201 and by myself on 8265 with no regressions seen.