Relax address availability check for multicast binds.
While it makes sense to limit bind(2) of unicast addresses that overlap
each other to be all from the same UID (like 0.0.0.0:53 and 127.0.0.1:53)
it makes little sense for multicast. Multicast is delivered to all sockets
that match so there is no risk of someone stealing traffic from someone
else.  This should hopefully help with mDNS as reported by robert@
OK deraadt@ bluhm@