refactor the code so that caching is setup in the repository that's linked to quirks during do_quirks.
then add a pre-cache that does run locate once on the full list of things
we want to update/install (assuming "regular" pkgsets, that is, where the
names are the ones we want).

Then the actual fetching of caching info will look in that pre-cache before
resorting to locate itself (or going straight to the wire).

The code isn't active yet, the XXX comment in
	OpenBSD/PackageRepository/Installed.pm
says it all.