Among published AT&T UNIX releases, a shutdown(8) command was contained in PWB/UNIX 1.0 (July 1977) and in System III (1982), but not in v7 (Jan 1979) nor in 32v (May 1979). According to the Australian UNIX User Group Newsletter Vol. I No. I https://www.tuhs.org/Archive/Documentation/AUUGN/AUUGN-V01.1.pdf , (October 1978), Letter from John Lions on page 20, UNIX/TS 1.0 (late 1978 or early 1979) also contained an implementation of shutdown(8). So it is unlikely it came from the CSRC and more likely it was invented either by the PWB or by the USG group. Since we don't know for sure, only cite the first public release.
It was a mistake to say "appeared in 4.0BSD".  That release only
contained a stub manual page (without a DESCRIPTION), and the
file man8/BUGS in 4.0BSD mentions it without giving details,
presumably trying to indicate it wasn't really available yet.
The is no shutdown(8) implementation in BSD before 4.1cBSD.

OK jsg@